# src/main/java/com/example/bankcards/dto/card/CardDtoBlockRequest.java
package com.example.bankcards.dto.card;


import jakarta.validation.constraints.NotNull;
import lombok.*;
import lombok.experimental.FieldDefaults;

@Getter
@Setter
@Builder
@NoArgsConstructor
@AllArgsConstructor
@FieldDefaults(level = AccessLevel.PRIVATE)
public class CardDtoBlockRequest {

    @NotNull
    String cardNumber;

}