# src/main/java/com/example/bankcards/dto/user/UserDtoOut.java
package com.example.bankcards.dto.user;

import lombok.*;
import lombok.experimental.FieldDefaults;

@Builder
@Getter
@NoArgsConstructor
@AllArgsConstructor
@Setter
@FieldDefaults(level = AccessLevel.PRIVATE)
public class UserDtoOut {

    String name;

    String email;

}